Extracts MyFixturesList into separate component
